Not compiling but at least project loads w/ sbt
paulp's sbt https://github.com/paulp/sbt-extras
works w/ fedora23 and it's much much easier to user reliably.
